Start code clean and better structure
